feat(merge): 폴더 정리
