+++ frontend/src/pages/Mainpage/Mainpage.css
