+++ frontend/src/pages/ProductPage/LoanPage.jsx
