+++ frontend/src/pages/ProductPage/assets/logos/black.png
