+++ frontend/src/pages/ProductPage/assets/logos/green.png
