+++ frontend/src/pages/ProductPage/assets/logos/hana.png
