+++ frontend/src/pages/ProductPage/assets/logos/index.js
