+++ frontend/src/pages/ProductPage/assets/logos/white.png
