+++ frontend/src/pages/ProductPage/components/BankSection.jsx
